Moved PixelMatrix to separate class
